move arg shape inspection to Alias command
