+++ crates/nu-protocol/src/return_value.rs
@@ -1,5 +1,6 @@
 use crate::hir::Block;
 use crate::value::Value;
+use crate::SyntaxShape;
 use nu_errors::ShellError;
 use nu_source::{b, DebugDocBuilder, PrettyDebug};
 use serde::{Deserialize, Serialize};
@@ -22,7 +23,7 @@ pub enum CommandAction {
     /// Enter the help shell, which allows exploring the help system
     EnterHelpShell(Value),
     /// Enter the help shell, which allows exploring the help system
-    AddAlias(String, Vec<String>, Block),
+    AddAlias(String, Vec<(String, SyntaxShape)>, Block),
     /// Go to the previous shell in the shell ring buffer
     PreviousShell,
     /// Go to the next shell in the shell ring buffer
